Publish javadocs via GitHub Action to GitHub Pages (#225)
Create action to publish pages and clean up build files slightly.
