+++ .github/workflows/javadocs.yml
@@ -1,0 +1,63 @@
+# Simple workflow for deploying static content to GitHub Pages
+name: Deploy Javadocs to Pages
+
+on:
+  # Runs on pushes targeting the default branch
+  push:
+    branches: ["main"]
+
+# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
+permissions:
+  contents: read
+  pages: write
+  id-token: write
+
+# Allow one concurrent deployment (if a build is already in progress, cancel in favor of the latest Javadocs)
+concurrency:
+  group: "pages"
+  cancel-in-progress: true
+
+jobs:
+  # Only run if the Java files were potentially updated
+  filter:
+    runs-on: ubuntu-latest
+    outputs:
+      should_skip: ${{ steps.skip_check.outputs.should_skip }}
+    steps:
+      - id: skip_check
+        uses: fkirc/skip-duplicate-actions@v5.3.0
+        with:
+          concurrent_skipping: 'same_content_newer'
+          paths: '["**/src/**", "**.gradle", "**/build.gradle"]'
+  javadocs:
+    # Check for potentially new Javadocs
+    needs: filter
+    if: ${{ needs.filter.outputs.should_skip != 'true' }}
+    # Set the URL to publish docs to (currently top-level GitHub Page for repo)
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
+    runs-on: ubuntu-latest
+    steps:
+      # Build Javadocs only
+      - name: Set up repository
+        uses: actions/checkout@v3
+      - name: Set up JDK
+        uses: actions/setup-java@v3
+        with:
+          java-version: '11'
+          distribution: 'corretto'
+      - name: Build with Gradle
+        uses: gradle/gradle-build-action@v2
+        with:
+          arguments: aggregateJavadoc
+      # Cache Javadocs
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@v1
+        with:
+          path: 'build/docs/aggregateJavadoc/'
+      # Publish Javadocs to URL
+      - name: Setup Pages
+        uses: actions/configure-pages@v3
+      - name: Deploy to GitHub Pages
+        uses: actions/deploy-pages@v1
