+++ c3r-sdk-core/src/main/java/com/amazonaws/c3r/config/EncryptConfig.java
@@ -72,7 +72,7 @@ public final class EncryptConfig extends Config {
         this.tempDir = tempDir;
         this.settings = settings;
         this.tableSchema = tableSchema;
-        transformers = initTransformers(secretKey, salt, settings, false);
+        transformers = Transformer.initTransformers(secretKey, salt, settings, false);
         validate();
     }
