+++ c3r-sdk-core/src/main/java/com/amazonaws/c3r/io/SqlRowReader.java
@@ -4,12 +4,12 @@
 package com.amazonaws.c3r.io;
 
 import com.amazonaws.c3r.config.ColumnHeader;
+import com.amazonaws.c3r.config.ColumnInsight;
 import com.amazonaws.c3r.config.ColumnSchema;
 import com.amazonaws.c3r.data.Row;
 import com.amazonaws.c3r.data.RowFactory;
 import com.amazonaws.c3r.data.Value;
 import com.amazonaws.c3r.exception.C3rRuntimeException;
-import com.amazonaws.c3r.internal.ColumnInsight;
 import com.amazonaws.c3r.internal.Nonce;
 import com.amazonaws.c3r.io.sql.SqlTable;
 import com.amazonaws.c3r.io.sql.TableGenerator;
