+++ c3r-sdk-core/src/test/java/com/amazonaws/c3r/action/RowMarshallerTest.java
@@ -9,6 +9,7 @@ import com.amazonaws.c3r.SealedTransformer;
 import com.amazonaws.c3r.Transformer;
 import com.amazonaws.c3r.config.ClientSettings;
 import com.amazonaws.c3r.config.ColumnHeader;
+import com.amazonaws.c3r.config.ColumnInsight;
 import com.amazonaws.c3r.config.ColumnSchema;
 import com.amazonaws.c3r.config.ColumnType;
 import com.amazonaws.c3r.config.DecryptConfig;
@@ -22,7 +23,6 @@ import com.amazonaws.c3r.data.Row;
 import com.amazonaws.c3r.encryption.Encryptor;
 import com.amazonaws.c3r.encryption.providers.SymmetricStaticProvider;
 import com.amazonaws.c3r.exception.C3rRuntimeException;
-import com.amazonaws.c3r.internal.ColumnInsight;
 import com.amazonaws.c3r.io.CsvRowReader;
 import com.amazonaws.c3r.io.CsvRowWriter;
 import com.amazonaws.c3r.io.CsvTestUtility;
