+++ c3r-sdk-core/src/test/java/com/amazonaws/c3r/io/SqlRowReaderTest.java
@@ -4,6 +4,7 @@
 package com.amazonaws.c3r.io;
 
 import com.amazonaws.c3r.config.ColumnHeader;
+import com.amazonaws.c3r.config.ColumnInsight;
 import com.amazonaws.c3r.config.ColumnSchema;
 import com.amazonaws.c3r.config.MappedTableSchema;
 import com.amazonaws.c3r.config.PadType;
@@ -12,7 +13,6 @@ import com.amazonaws.c3r.data.CsvRowFactory;
 import com.amazonaws.c3r.data.CsvValue;
 import com.amazonaws.c3r.data.Row;
 import com.amazonaws.c3r.exception.C3rRuntimeException;
-import com.amazonaws.c3r.internal.ColumnInsight;
 import com.amazonaws.c3r.io.sql.SqlTable;
 import com.amazonaws.c3r.io.sql.TableGenerator;
 import com.amazonaws.c3r.utils.FileUtil;
