+++ c3r-sdk-core/src/test/java/com/amazonaws/c3r/json/ColumnTypeTypeAdapterTest.java
