+++ c3r-sdk-examples/src/main/resources/log4j2.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+  ~ SPDX-License-Identifier: Apache-2.0
+  -->
+
+<Configuration status="INFO">
+    <Properties>
+        <Property name="log-path">logs</Property>
+    </Properties>
+    <Appenders>
+        <RollingFile name="APPLICATION" fileName="${log-path}/application.log"
+                     filePattern="logs/$${date:yyyy-MM}/app-%d{MM-dd-yyyy}-%i.log.gz">
+            <PatternLayout>
+                <pattern>%d{dd MMM yyyy HH:mm:ss,SSS} [%p] (%t) %c: %m%n%ex</pattern>
+            </PatternLayout>
+            <Policies>
+                <OnStartupTriggeringPolicy/>
+                <TimeBasedTriggeringPolicy/>
+                <SizeBasedTriggeringPolicy size="50 MB"/>
+            </Policies>
+            <DefaultRolloverStrategy max="20"/>
+        </RollingFile>
+        <Console name="CONSOLE" target="SYSTEM_OUT">
+            <PatternLayout>
+                <pattern>%d{dd MMM yyyy HH:mm:ss,SSS} [%p] (%t) %c: %m%n%ex</pattern>
+            </PatternLayout>
+        </Console>
+    </Appenders>
+    <Loggers>
+        <Root level="INFO">
+            <AppenderRef ref="APPLICATION"/>
+            <AppenderRef ref="CONSOLE"/>
+        </Root>
+        <!-- Disables wire logging -->
+        <Logger name="WIRE" level="OFF"/>
+        <Logger name="org.apache.hadoop" level="OFF"/>
+        <Logger name="org.apache.spark" level="WARN"/>
+        <Logger name="org.sparkproject.jetty" level="WARN"/>
+    </Loggers>
+</Configuration>
