+++ c3r-sdk-parquet/src/main/java/com/amazonaws/c3r/action/package-info.java
@@ -2,9 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /**
- * This package classes used to marshall (encrypt) and unmarshall (decrypt) data to and from the clean room for Parquet files.
- * {@link com.amazonaws.c3r.action.RowMarshaller} handles the logic of marshalling data outside of anything having to do with
- * Parquet itself and {@link com.amazonaws.c3r.action.RowUnmarshaller} does the same for unmarshalling.
+ * This package classes used to marshall (encrypt) and unmarshall (decrypt) data to and from the clean room for the various supported
+ * data types. {@link com.amazonaws.c3r.action.RowMarshaller} handles the logic of marshalling data outside of anything having to do with
+ * the actual data format and {@link com.amazonaws.c3r.action.RowUnmarshaller} does the same for unmarshalling. Each format specific class
+ * handles file I/O and value creation only for that particular data type.
  *
  * <p>
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
